Added HInitializer class to encapsulate class initializer methods the same way HConstructor encapsulates class constructors.
Also, some minor javadoc editing.

Lastly, changed the way ImplMagic constructor HMethods so that the HCodes
are all generated on demand, instead of at once on loading.  Should still
garbage collect the original classfile structure correctly, once all the
method HCodes are generated.  This may speed up class loading slightly.

There is not yet an HInitializerSyn class to accompany the HInitializer.
I'm planning on majorly overhauling the HClass/HClassSyn system anyway,
so this isn't a priority for me right now.  Nevertheless, it's on the
TODO list.