The day of immutability has arrived.
This commit changes the access modifier on the fields of all quads to
'protected' to enforce immutability.  New accessor methods are used to
query the values of the fields.  Some mutability still remains, especially
with respect to the temp renaming interface, but this will eventually
disappear.  I have modified all the code I could find on the magic-2-0
branch to work with the immutable quads.  Darko's stuff should work fine.
Since Nate and Mark's stuff is not current on magic-2-0, the changes I
have checked in for their code should be viewed as guides.

Please examine the change to CALL carefully.  The sense of the final
boolean parameter to CALL has been inverted, because 'isSpecial' didn't
seem to mean anything much.  The flag is now 'isVirtual' and is usually
true.  It is false for: constructors, static methods, private methods,
and code which uses the 'super.' prefix in the java source.  See the
JVM reference entry on 'invokespecial' for more details.

Also, METHODHEADER has now been officially put to rest.  There is just
HEADER now.  Again, I have modified the code to work with this change; you
may want to double-check code of yours that I have modified to be sure
I did the right thing.

There is still much room for tweaks and improvement, but this is the
first big step towards immutable quads.  Please offer your comments.

Oh -- and there's lots of new javadoc for the Quads now, too.  I also
added lots of assertions to ensure that all created quads are valid.
Since you can't change quads after you create them, they better be
right from the start.

Have fun & let me know if I broke anything.