Added harpoon.IR.Quads.QuadFactory and harpoon.Temp.TempFactory classes to generate consistent and repeatably numbered Quads and Temps.  This numbering should be repeatably across runs of the program and insensitive to the order in which methods are translated to bytecode.
Also split the HEADER quad into HEADER and METHOD so that the HEADER node
never needs to be modified or replaced.  Also discarded the 'footer' field
of HEADER, replacing it with an explicit link, which simplifies things
considerably and Just Makes Sense.  See the javadoc for details.

The HANDLER quad, which represents explicit try-catch blocks, has also
been committed, although the current translator doesn't generate them.
The next generation translator, which is not ready to check in yet, will
generate HANDLERs and ARRAYINITs so start thinking about them.

I have updated all code in the magic/2.0 branch in accordance with these
changes.  Please double-check my work and make sure I haven't missed
anything.

Nate will be glad to know that you can finally get the Method and HCode
containing a Quad from the reference to the Quad (via the QuadFactory).
Yay.