Just got back from South Carolina today, so I've got lots of code to commit:
Added a TYPECAST quad to mark type casting of temporaries.  TYPECAST is
a subclass of NOP, so existing code requires no changes.  Type analysis
can take advantage of the extra information, and the implicit-exception form
is cleaner with the TYPECAST quads.

Normalized the COMPONENTOF and INSTANCEOF quads.  They both work on
null-pointers now -- before, only COMPONENTOF was supposed to work on null,
and INSTANCEOF was supposed to be used only on quads that had been previously
determined to be non-null.  That silly restriction is gone.  Some optimizations
may work better if the explicit null-check before INSTANCEOF (which is now
absent) is re-inserted.  It's on the TODO list.

MAJOR FIX to harpoon.ClassFile.Loader: through a slip of my fingers, the
Loader class was actually treating all non-zip file entries as '.' (ie,
ignoring the directory portion of the path).  I don't think anyone noticed
this, but they were bound to eventually.  I kissed it and made it better.

Many improvements to harpoon.Interpret.Quads classes.   Interpreting
harpoon.Main.Main now works.  Implemented more native methods
(INFileInputStream/INObject/INRuntime/INFloatDouble) and removed some
java language reflection in harpoon.IR.Quads.OPER to make this possible.
[try java harpoon.Main.Run harpoon.Main.Main harpoon.Main.Main.... and
 be prepared to wait a while.  I give java the -noasyncgc and -mx70m flags
 when I do this.]

Profiling support in harpoon.Interpret.Quads.  The -prof option works, with
syntax identical to java:
 java harpoon.Main.Run -prof:prof.out [classname] [args]
puts profiling information in prof.out.  All methods invoked and objects
created are logged.  Separate tools (not included in this commit) analyze
the data.  The profile data is output in a simple plaintext format, but it
is gzip-compressed on the fly to save space.  Use gunzip or zcat if you want
to take a gander.

Added harpoon.ClassFile.HClass.isInstanceOf() method which supersedes the
HClass.isSuperclassOf() method.  Several instance of tests in the codebase
were using isSuperclass which incorrectly reports interface inheritance and
array hierarchies.  The new isInstanceOf() method correctly reports if
one class or array is an instance of another class, interface, or array type.
Affected: harpoon.Analysis.QuadSSA.SCC.SCCAnalysis
          harpoon.Interpret.Quads.{Method,INSystem}
[also some minor tweaking to harpoon.ClassFile.HClass to cache class constants]

Miscellaneous:
 Updated TODO list.
 Improved assertions in IR.Quads.HANDLER to ensure that the caught exception
  is a subclass of java.lang.Throwable (or null, which means "any").
 As mentioned, use of java language reflection in IR.Quads.OPER is removed;
  this entails adding a two methods to IR.Quads.Qop.
 I commented out the proxy quad construction methods in IR.Quads.QuadFactory,
  since I don't think they are a good idea.
 Support for profiling in Main/Run
 Added WriteableTempMap interface (subclass of TempMap) which will be used
  by some of the translation routines.
 Logged 1400+ miles in my pickup without initiating unplanned rotation,
  revolution, or convulsion of the vehicle.  Had to replace my windshield
  wiper fluid because it froze solid in Boston.  Finally thawed in the
  Carolinas.