Wrote interfaces for all the various numberings needed.
MethodMaps map methods to ordered numbers.  Felix, your interface numbering
scheme should implement a MethodMap to export its results.
There are also FieldMaps to order fields, ClassDepthMaps to report
on class nesting depth, and InlineMaps to determine whether or not
a field can be inlined.  FinalMaps determine whether a class, method, or
field is final.

Some of these maps have simple implementations that return conservative (if
not-too-bright) results; some are non-trivial.  If you are working on the
back end, you should closely examine this map structure to determine which
maps your code should be implementing.

At this point these are all abstract superclasses rather than interfaces.
I would like to keep it this way, both for efficiency and because I would
rather the code for each mapping function be kept separate, rather than
combined in some huge unmaintainable map-of-all-trades.  If you can justify
a reason why a particular map should be an interface instead, talk to me.