After a bit of experimentation and thought, I decided to put an abstract declaration of equals(Object) in but not hashCode().
My reasoning is this: the default equals() implementation of Object
is easily duplicated with ==.  Therefore we don't lose the functionality
of the method by abstracting it in the ColorableNode class.  However,
the default hashCode() implementation is NOT easily duplicated, or
at least I don't see how.  I considered adding a method objectsHashCode()
to ColorableNode, that would allow subclasses access to the implementation,
but then I realized that such complexity was counterproductive.  So
right now subclasses of ColorableNode are forced to implement equals()
but it is the responsibility of the programmer to see when it is
necessary to override the default hashCode() implementation.