Added a group of classes to replace the UseDef class.
TempToHceArrayMap seemed like a logical addition (a
superclass containing common code for both the UseMap
and the DefMap) but because the logical names for the
methods offered by the two objects differ, I ended up
still doing a fair amount of unnecessary code duplication
between the two Maps.  Any suggestions on how to get
around this?