1. Modified doc for Worklist to conform to requires/modifies/effects style.  2. Implemented redudant Worklist functionality in Set (means that extensions of Set have less to implement, though I did add an abstract getArbitrary() method to handle the pull() method of Worklist... 3. Removed reimplementations of Worklist methods from HashSet (since they are now implemented in Set) and added implementation of getArbitary.
I have also begun work on a TreeSet class (a ListSet couldn't
work, at least not efficiently, because we ensure that the
addition of duplicates is not allowed.  THis implies that a
sorted storage mechanism is useful (thus a Tree Structured Set,
so that we don't have the overhead of Hashtables).  Will
continue work on it later.