Found that requiring an HCode for these Maps was too strigent because I couldn't properly perform analysis on clones of codebodies. Looking through the code, I found that the heart of the matter was that I needed an arrayFactory, not the HCode itself (it was just simpler to pass an HCode into the constructor.)  So I changed the constructor to TempToHceArrayMap so that takes an arrayFactory.
Then I added a constructor to UseMap that takes an arrayFactory and
a HCodeElement[], to accomidate the analysis needed as mentioned above.
A similar constructor could be added to DefMap easily, but I didn't
bother to.

(NOTE: Should we possibly add getArrayFactory()
to the HCodeElement interface, so that classes that need to duplicate
HCodeElements into new arrays can do it properly without this sort
of hackery?)