Added support for THROW quads.  The MethodInliningCodeFactory should now work for general byte code.
Also, per Scott's request, added code to clone the HCode prior
to having the inliner do anything to it, to minimize propagation
of effects across optimizations interfering with eachother.  Scott,
will cloning HCode do everything you want me to?  IE, will it
also produce a new QuadFactory for the HCode to use?  If not, can
you point me in the right direction as to how to do what you want?