Forgot to add this (bad Scott).
The Collections API makes you implement the Map.Entry Interface everytime
you extend AbstractMap.  There are only two "interesting" methods in
Map.Entry, getKey() and getValue() -- the rest is code which will almost
always be the same.  So we provide an abstract class implementing Map.Entry
to take care of the grunge work.