Checking in all files for the tree interpreter.  These files fall into 3 categories:
1)  Classes which were not present in the quad interpreter.  These files exist
primarily to provide an implementation of pointers.  These files include:

ArrayPointer.java           // Pointer to array reference
ClazPointer.java            // Pointer to static class data
ConstPointer.java           // Points to method or static field
FieldPointer.java           // Points to non-static field
InterfaceListPointer.java   // Points to interface list in static class data
Pointer.java                // Abstract Pointer class
StringPointer.java          // Pointer to String constant
UndefinedPointer.java       // Pointer to data for which the type is not known

PointerTypeChangedException.java  // Thrown when one Pointer is converted
                                  // to another Pointer class

Additionally, a specialized OffsetMap and AllocationStrategy were added:

InterpreterAllocationStrategy.java  // Strategy specific to Tree Interpreter
InterpreterOffsetMap.java           // OffsetMap specific to Tree Interpreter


2)  Classes which were modified from the quad interpreter.  These include
ObjectRef and ArrayRef, which were modified to allow access to their data
through the Pointer classes.  However, I tried to keep changes to these classes
to a minimum to preserve their compatibility with the native method classes.
Substantial changes of course were made to the Method class, which is the
bulk of the Tree Interpreter.  Changes were made to the class loader in
StaticState to make it possible to access static data through the Pointer
classes.  Finally, TreeStackFrame is a direct knock-off of QuadStackFrame,
except that it uses the Tree classes.


3)  Classes which are unchanged from the Quad interpreter (everything else).


This information will be duplicated in a README which I'll leave in this
directory.