The InstrMEM class is an abstract way to represent memory operations in assembly (needed to add Loads and Stores after register allocation)
The InstrVisitor class is part of the continuing Visitor Pattern.
I'm adding it because the Solver from the DataFlow package uses
the Visitor pattern in all of its calls...still not sure if this is
the right way, since we don't have a Visitor superclass that all
the other visitors extend, and we probably need that to make a truly
generic DataFlow solver.