1. Added DEBUG flags to much of the DataFlow package in an effort to track down a bug in LiveVars
2. Added a few more methods to InstrSolver to let me initialize
the Worklist with more than just one Object

3. Fixed bug in usage of LiveVars (basically I wasn't initalizing the
Worklist with enough elements; see (2.) above)

4. Got Local Register working

Should look into making the Solver stuff generalized.  There's
no need for an actual InstrSolver (it doesn't internally
depend on the BasicBlocks being composed of Instrs, it just
tries to enforce that constraint on the Visitors you pass
in).  QuadSolver might be able to be generalized into a Solver
that does generic DataFlow analysis, given an code for an IR and
an appropriate Visitor for that IR (or for Properties of that
IR; LiveVars doesn't require Instrs, just UseDefs I believe)