It turns out that MOVEs need to be typed, too, since moves between (for example) floating-point registers, or 64-bit long values, are different than moves between ordinary 32-bit registers (on a 32-bit architecture).
The type of a move is the type of its src and dst expressions.  There's
an assert() in IR.Tree.MOVE to guarantee that the src and dst types
always match.  Use a Tree.UNOP to coerce the types if you're trying to
make an assignment between, say, pointer and integer.

The grammar and Tools.PatMat.Spec have been updated to support typed MOVEs.
See the upcoming commit to Backend/Sparc/CodeGen.spec for an example of
where typed MOVEs are necessary.