Not content with fame in the on-line world, C. Scott Ananian attempts to defy fate's allotment of *only* 15 minutes of global publicity in commiting Yet Another Brilliant Piece of Code to the revolutionary FLEX project.
Except in this case, the stellar patch only adds support for Tree DATA
elements to the code-generator-generator (i love the sound of that)
parser/lexer --- which a properly-trained monkey could have pounded
out, given enough time.  If only animal-training were easier!
But I can't stand the urine, the newspaper, and the metal bars...

http://www.salon.com/tech/feature/1999/07/30/redhat_shares/index.html
 ^
  above URL added for the benefit of the ChangeLog. ;-)