Modified ToTree to generate CALLs such that
  *  CALL.retex is actually a pointer to exception handling code,
     as opposed to a location for exceptional return values
  *  There is no code to explicitly check for a thrown exception after
     a procedure call

I also seized this chance to clean up a few things in the ToTree class
that were bothering me:

  * Collection-ized the code
  * Fixed a bug which would generate correct (but foolish code) that would
    insert redundant JUMP instructions that would simply JUMP to the next
    statement
  * Added a few comments & made some of the stranger sections of code a bit
    more readable

Sorry Scott, I didn't modify the THROW object to contain a pointer to the
exception handler.  I couldn't think of a way to do it that wouldn't have
been more trouble than it was worth, and the code generator doesn't really
need this information.

I still have to add code that deals with the case when my simplistic
pattern matcher doesn't match the explicit exception checks generated by
the quad form.