The first in a long series of commits which replace the old TypeMap interface. The old typemap interface was:
	HClass typeMap(HCode, Temp);

This was fine for SSA form, but in non-SSA form, it is possible that
the same temp may be assigned values of different types at different points
in the control flow graph.  The new interface accounts for this:

	HClass typeMap(HCodeElement, Temp);

I've update all of the code affected by this, except for Brian's.  Sorry dude,
I would've had to make changes that I didn't feel justified in making to
unfamiliar code, so I'll leave it to you.