Reorganization of HData interface: Tree.Data no longer extends Tree.Code.
The Tree elements valid in an HData (and only those) have been made
to implement HDataElement, which is more a 'property' than an interface.

The TreeFactory has been split up into IR.Tree.Code.TreeFactory and
IR.Tree.Data.TreeFactory so that we don't have methods which are not
valid for 'HData's in TreeFactory.