Fixed up mozilla hackery so that cvsblame.pl works like it was supposed to.
This is useful even w/o the WWW hookup; type
  bin/cvsblame.pl [filename]
to get a plain-text annotation of who changed what in the file.
The -h option gives a usage message.

Unfortunately, some of the mozilla hackery seems to have broken
cvsblame's ability to parse the CVS/Root file to find out where
the version control information is found; and it is not certain
that it would have dealt with remote respositories and CVS_RSH
correctly even originally.  So, for the moment you need to
specify the complete path to the CVSROOT copy of the file; that
is,
  bin/cvsblame.pl ~cananian/Harpoon/CVSROOT/Code/[filename],v
...which means you have to be on the machine where CVSROOT is
located (lesser-magoo) or have NFS access (which nobody does).
I'll fix this RSN, but you can play with it in the meantime
regardless.

Oh -- and check out
 http://flexc.lcs.mit.edu/Harpoon/cvsweb.cgi
too -- WWW browsable CVS logs with readable color-coded diffs between
arbitrary revisions.  I'm working on integrating this into the
funky Java syntax-highlighter hyper-linking annotation stuff, and
tying cvsblame together with both.  Fun, fun, fun.