Fix infinite bug-a-boo in ToNoSSA.  I think this fixes the problem we were seeing in IR.Tree, too.
Quad.replace no longer "magically" updates the HandlerSets, since we can't
always do this reliably.  Now there's a separate 'transferHandlers'
method of IR.Quads.Quad which does this.  The only transformation which
required this functionality was Quads.Peephole; I've updated it appropriately.

ToNoSSA now uses the general Quad.replace method instead of reimplementing
it (incorrectly) itself --- Quad.replace is a little trickier than you
might assume because of the possibility that an edge both comes from and
goes to a single quad.  This is something you are only likely to see in
certain very-simple-and-usually-infinite looping constructs.  Please
check your code, if you've reimplemented Quad.replace at some point, and
replace your version with a call to the canonical version.  Thank you.

We now return to our regularly scheduled Monday night.