Big semantic change:  INSTANCEOF and COMPONENTOF no longer are valid when applied to null pointers, except in quad-with-try form.
Also, a field has been added to ASET so that we can avoid generating
COMPONENTOF tests when the array components are of primitive type
(program typesafety guarantees COMPONENTOF would return true in this
case).

Finally, some optimization was done to the translation of bytecode
test and branch instructions: the test instructions were changed to
return type 'int' (instead of type boolean) and the bytecode instructions
'ifeq' and 'ifne' no longer need to do an explicit comparison against
zero; the CJMP actually performs this function.  The translation of
the bytecode INSTANCEOF operatior is much improved.

Changes were made to Quads.Translate, Quads.UnHandler, and the Quad
interpreter to support the semantic changes.  Ran some tests on the
system and verified that everything still seems to work correctly.