Improve code emitted to load constants.
We now load longs, floats, and doubles correctly (not just ints)
and output a pretty decent approximation to the minimum number of
instructions required in every case.  Nifty feature is that the
code is smart enough to load the complement of the number, if
the number of ones in the constant is larger than the number of zeros.
This means, for example, that loading "-4" takes one cycle instead of four
now.

Groovy.