Committing my evil evil register allocation hacks to the repository so folks can look at them / use them until we get rid of them.
This check-in is the original source I wrote for Appel's class.
It follows the APIs in his book, and so it might be more useful
reference if you're trying to figure out how his book structures
register allocation.

It won't compile.  The next commit will be the nasty evil changes
which make this existing code work with FLEX.  So hold off on the
'make update' for a couple secs, ok?