Make note about current bug in CALL instructions, probably likely to manifest itself only under high register pressure.
Optimize away stack size re-adjustment in callEpilogue in the case that
no adjustment is needed.

Add patterns for add of register and 8-bit immediate.  Lots of these
patterns are needed, for every instruction which takes an operand2,
but I think this is a job for a code generator generator generator.
[or is it a code generator specification generator?]