Makefile mods to get Sparc stuff building, to oblige andyb.
The rest of you probably don't want to 'make update' until andyb's gotten
the SPARC stuff compiling --- this commit breaks the build until that
happens.

You should be able to add the line
 Backend/Sparc
to Harpoon/Code/.ignore
to ignore the broken Sparc stuff, but I'm not sure that will manage to
ignore the broken .spec file.  So just stick with the old GNUmakefile
for a while, eh?