Ack.  I hate screwy architectures.
Both of my previous commits were partially wrong, because the sense of
the comparison test is *inverted* for half of the comparison operations.
So I only got half the tests backwards last time.  Of course, none of this
is documented anywhere, so I had to mess around with little assembly
programs to figure it out.

Here's the deal:  ___ne*, ___gt*, and ___lt* return the result you'd expect:
non-zero for true, zero for false.  However, ___eq*, ___ge*, and ___lt*
return an *inverted* result: non-zero for false, zero for true.

The new, improved spec file should do the right thing, now.  One hopes.