Scott's magical perl script.
This little beauty zips through the FLEX source files and automatically
replaces import * statements with the proper list of classes. [Actually,
it generates a diff which you then pipe to the 'patch' program to make
the changes.]  This means... (wait for it...) I'm not going to yell at
anybody for using import * anymore!  Use them all you like!  I'll just
run my magic script once in a while to clean everything up.

This script uses the mysterious '+M' option of jikes to aid in its
wizardry.  Just thought you might be curious about that.