Updated harpoon.IR.Tree.Code and subclasses to return TreeDerivation objects instead of implementing Derivation.  Cloning does not yet preserve TreeDerivation information, as the Tree.clone() interface is broken.
ToCanonicalTree has also been updated to use TreeDerivation and
the associated DerivationGenerator class, but as I was updating I
noticed that derivation information for the canonical form was
never actually being generated (even though all the hooks were
there).  So the updated ToCanonicalTree has the same lack of
functionality as its predecessor.  This too will change...