Fixed bugs in Tree.AlgebraicSimplification.  First off, in Java (and in C, too, for that matter) the bitwise-and operator '&' has higher precedence than the bitwise-or '|' operator.  This means that expressions like    (somevalue & mask1|mask2|mask3) != 0 will *always* evaluate to TRUE.  AlgebraicSimplification had three instances of this
Along the same lines, there was an expression of the form
   (value & ~(mask1|mask2|mask3)) != 0
Now, while the *intent* of this expression was obvious from the context
(to ensure that 'value' does *not* have any of the bits set in mask1,
mask2, or mask3) the way it was written was incorrect.  Bitfields are
tricky.

There was one instance of BINOP.operandType() being used where BINOP.type()
should have been used.  Felix and I have talked before about how confusing
these two methods are, but the way we have things makes sense if you think
hard about it.  type() always returns the *result* type of the expression,
which for certain operations (comparisons, for example) is unrelated to the
*operand* type.

Lastly, there was a weird instance of 'UNOP.optype' being used where
'UNOP.op' should have been used.  Very odd.

AlgebraicSimplification now works on the canonical 'Hello, world' example
(it executes without errors in StrongARM native code) and I'm pretty
confident it will work for larger cases as well.