Lots of changes to fix the two bugs I have found into Martin's algorithm last week:
 1. we need to be *more* conservative in the inter-procedural analysis.
In particular, at the analyzed CALL sites, more nodes from the callee's
graph should be introduced into the graph of the caller. The rules given
in the paper are insuficient. Fix: all the callee's nodes are initially
inserted; after the recomputation of the escape info, some epty loads and
"fake" outside edges are removed. This way, the generation of the new graph
is separated from its simplification. Status: DONE

 2. the outside edges must be mapped not only against the inside edges
from the other scope but also against the inside edges from their own scope.
Fix: add an edge ordering structure to record the inside edges created before
a specific load node is read. Only these inside edges are used (using all the
inside edges will disrupt the performances). Status: the data structures are
done, the edge ordering is generated. I still have to use it in the
inter-thread part of the analysis.