Adding a new functionality suggested by John Whaley: - a method to obtain the parallel interaction graph valid at an arbitrary point of the program and strngly connected to it: - a method to obtain the set of the nodes pointed by a certain Temp (variable) right before executing a certain Quad in the body of of a (meta)method.
Of course, all this info was computed even in the previous
version (since our analysis is flow sensitive), what was
needed was a convenient interface and some hacks to select
the important things.

Thsi will be useful in finding exactly what nodes are
pointed by the variable appearing in a sync operation.