I added a more compact version of the basic blocks. We use too many Sets, Maps etc. They are all very heavyweight structures: not only they take lots of memory but their traversal is extremely expensive: to traverse the instructions from a bsic block, one need to obtain a List, next an Iterator (a very complex dynamically allocated object) over that List and next do a for loop where the condition is it.hasNext(). This is a few orders of magnitude more expensive than storing everything in an array and doing an int-indexed loop. Moreover, the average size of a basic block is somehing between 2 and 3 instructions: having a Set for one element and an Iterator to grab is catastrofic ...
The LightBasicBlocks are array based. They are just
a more compact way of storing the BasicBlock view
of the code. In particluar, there is direct pass
from HCode to LighBasicBlocks. You need first to
convert the HCode to classic BasicBlocks and these
are next converted into LightBasicBlocks.

As a general advice to everybody, consider the
Java Collections as convenient data structures
useful while generating some info, but too
expensive for storing that info.