./configure now defines _REENTRANT for proper C-library operation when heavy-weight threading is configured.
Added support for pthread_rwlock_t to configure.in.  It is part of the
UNIX98 standard, but some pthread implementations don't have it yet. We
fall back to a mutex if we have to.  glibc needs _XOPEN_SOURCE=500 defined
in order to pull in the UNIX98 stuff, so we test for that case, too.