Array clone methods are problematic: FLEX keeps them separate from java.lang.Object.clone() because Object.clone() can throw exceptions while <array>.clone() is guaranteed *not* to throw an exception... the actual implementation of the two is rather different, too.
But java bytecode conflates the two, which means that our input routines
can not reliably distinguish between a call to Object.clone() and a
call to <array>.clone() w/o at least a type analysis.  We hack around this
for the time being by adding all <array>.clone() methods to the callable
methods set.  Another solution might be to eliminate the separate
<array>.clone() method -- just use Object.clone() for both -- which means
that the runtime's implementation of Object.clone() has to be fixed to
dispatch the proper array clone method when appropriate.
For primitive arrays this gets quite ugly.