1. Modifications in the test program to show info only about the methods reacahable (i.e. callable) from "main", ignoring those that are called only by the JVM (before starting main). This eliminates some unuseful messages.
2. In the inter-procedural analysis, if a call site
has no callee, it means it never appears in practice
(some classes are not instatiated: a good example is
SecurityManager - if the programmer doesn't set up any
(and none is set up by the JVM (or browser), no
method from that class can be called). So, to conclude,
the call should be simply ignored, instead of
considering it a hole as before. Of course, this rely
on the supposition that ClassHierarchy and
(Meta)CallGraph are not buggy.