Lots, lots of changes:
1. rewrite of the PAEscapeFunc class. It turns out that we
are not concerned about the specific call site (to a native
method) where a node escapes - all we need is to know if
it escapes through a method or not. This is because once
a node escapes into a native method, it is escaped forever.
This new version should reduce somehow the memory
consumption.

2. Some native methods are treated specially instead of
just being considered holes: java.lang.Systsem.arraycopy,
java.lang.Throwable.fillInStackTrace and
java.lang.Thread.setPriority0(). Others to come ...