Here is some new version of the pointer analysis stuff. Some modifications were made:
1. the edge ordering relation turned out to be really
expensive to maintain and not very useful (for recursive
methods or methods with loops, it seems to become OxI).
It has been dumped over the bord! (everything is
controlled by the PointerAnalysis.IGNORE_EO flag)

2. When a node reached the upper limit of the
specialization chain, it doesn't contain precise information
about the call chain that created it. For example, if
that limit is 2, nodes on level 0 represent objs allocated
in the current method, nodes on level 1 represent obj
allocated in the immediate callees and nodes on level 2
represent objects allocated on call path of 2 or more
CALLs. Since the call path info is not precise, instead
of generating one specialized node for each call site,
I just generate a single one (for all the call sites) called
"bottom". This happens *only* for nodes at the edge
of the specialization limit (who were anyway impossible
to use into optimizations), the other nodes are not affected
This is siupposed to reduce the number of unuseful
specializations.

3. Some aggressive shrinking strategy has been adopted
for graphs: at the end of a method, after retaining only
that part of the graph which is reachable from the
outside, the LOAD nodes which don't have any sync action or
inside edge in teh tree rooted in them are eliminated.
This is not good for some PA uses (for example, if we
just try to determine how far into the memory is one
application reading) but is OK for us. Controlled by
the ParIntGraph.AGGRESSIVE_SHRINKING flag. Supposed to
reduce the quantity of outside edges and outside nodes that
we carry from one method to another (especially in nests
of mutually recursive methods).

4. A simpel type inference package was added to detect the
AGET operations on arrays of primitive types (eg int[],
char[]) These arrays are not relevant for the PA; this way
we eliminate some potential load nodes.