Updated Lex.jar binary to JLex version 1.2.5b6, built from seperate sources for better line number information.
This is just so that the native JLex binaries I'm currently building
have line number information which correspond to some version we actually
have sources for.  The previously-checked in JLex binary was for some
intermediate revision between 1.2.4 and 1.2.5 which was "current" on
my development branch at the time I build the class files in Mar 1999.