The method inlining transformation for improving the effectiveness of the stack allocation is pretty much done. (I tested it on barnes.par.java and the results are VERY good).
Features:

1. The method inliner is not dummy. In particular,
we don't inline all the methods before the analysis
and wait to see what happens; only call sites that
can increase the stack allocation are actually
inlined.

2. The order into which the different CALLs are
inlined is important: if m2 should be inlined into
m1 and m3 should be inlined into m2, it is clearly
better to do "m3 into m2" first and "m2 into m1"
next (it is not incorrect to do it the other way but
the resulting code is less efficient since m1 will
include the original, non-optimized version of m2).
The inliner embedded in MAInfo.java achieves this
by doing a topological sort of a component graph
and processing the inlining requests in reverse
topological order.

3. The inliner supports circular inlining hints, eg
"inline m1 into m2", "inline m2 into m3" and
"inline m3 into m1". However, the results are not
optimal in this case; this case is treated just to
eliminate the need for testing against it in the
"inlining hints generator".

4. For the time being, we do only 1 level inlining,
ie the analysis gives hints like "inline this CALL".
However, it is possible to modify it to treat more
difficult cases such as "inline the call chain
CALL1, CALL2, ... CALLk"