Purpose statement for the 'Tests' module:
This cvs module contains all sorts of test suites, example programs,
benchmarks, and etc. that might be useful to the FLEX project.

Please move anything in Code/Test to this module; any benchmark programs
you are using but have not checked in should also be added to 'Tests'.

Note that this module is "Harpoon/Tests" as opposed to the current
"Harpoon/Code/Test" directory.