Globalref stack is now stored as a doubly-linked list, which makes adding and deleting references both constant-time.
Localref stack is now stored as a contiguous stack *not* as a linked-list.
Default stack size is 64k; this may need to be tweaked.  This makes
adding/deleting refs quicker, and makes DeleteLocalRefsUpTo() constant-time.

include/precisec.h now has implementations for PUSHOBJ() POPOBJ() and such
to fully support precise garbage collection in C.