Reorganized statistics-collecting/printing code.  Decoupled it from the clheap stuff.
There is a new src/stats module for statistics-related code -- only
stats.c for now.

There is a --with-statistics option to ./configure to enable statistics,
and a WITH_STATISTICS #define to test within code.  Usually this won't
be necessary, though, as include/fni-stats.h will #define the
DECLARE_STATS/INCREMENT_STATS/etc macros to nothing if WITH_STATISTICS
is not defined.  This cuts down on the number of disruptive #ifdefs
needed inside functions.

All statistics are declared locally in src/stats/stats.c.  Some changes
to src/clheap/*.c have been required to update them to the new statistics
interface.