Let a stack object have its layout be determined by a METHOD as well as an INVOCATION.  This allows use of the stack object by the parameter loading code, which has access to a METHOD.
I think it is kind of ugly the way METHODs and INVOCATIONs are
different, and I had to duplicate code for dealing with METHODs
because they do funny things like store the exceptional return address
as the 0th parameter.

I believe the best solution (which would allow us to have a virtual
frame pointer (oh, except for exception handling)) to this problem
would be to put in psuedo-insts that get fixed up in procFixup.  At
that point our regular stack object knows enough about the frame
layout to be able to make parameter loads stack relative instead of FP
relative, which is what they are now.