1. Change how float and double constants are dealt with.  I actually emit a   1: .float x   lw dst 1b
This might be unnecessary, but I figured Float.floatToIntBits had its
own algorithm for representing floats that would be machine indpendent
and not necessarily match the mips representation.  Is this right?

2. Some stack stores incorrectly list sp as a destination.

3. More damn fallout from the mipspro debacle.
What happens when you change .balign to .align?  Well, during startup
when the runtime is building the hashtable of interned strings, it
steps through the string_constant section with a stride of
sizeof(string class).  Well, we find lots of ZEROs because we aligned
to 4 words instead of 4 bytes.  Darn.