Fix a nasty & subtle bug with long integer comparisons.
Basically, I copied some code from the StrongARM architecture
reference manual without noticing that it was only valid for
*unsigned* 64-bit values.  For signed values, things are quite
a bit more complicated.

The current implementation uses 6 instructions (one of which is a
branch) for all comparisons except for equality/inequality, which use
only 4 instructions (none of which are branches).  It may be possible
to improve this number -- we'd especially like to remove the branch
and its associated pipeline delay.  Perhaps the brilliant insight
on how to do this will come to one of us one day.

In any case, we generate nicer code for these cases than gcc does
at the moment.  That should be enough to satisfy anyone, right?