Added '-r [root set file]' option to Main/SAMain to allow adding additional classes/methods to the root set on the command line.
Also checked in sample Support/rmi-root-set file to show how
it's done.  The file format is one line per class/method,
with an 'include [filename]' directive also supported.
Methods need to be suffixed with their descriptor in order
to resolve overloading.  A method which takes a string
would be listed as:
  package.classname.method(Ljava/lang/String;)V