The AllocationInformation has been extended with a new flag that indicates whether the synchronizations on objects allocated at that NEW/ANEW site are useless or not. The allocation native method is supposed to save this flag into the allocated objects so that the synchronization native method can avoid doing the work for the marked objects.
Correspondingly, MAInfo was extended to generate this
information. Currently, if the memory allocation info is
generated, the synchronization info is generated too. In
the future, it would make more sense to separate them and
have different controlling flags in harpoon.Main.PAMain

Finally, guess what?, none of the new features was tested in
any way ...